requirements.txt
streamlit
gtts
langdetect
pytesseract
pdf2image
pillow
openai-whisper
